refactor : attributes의 위치를 Request에서 책임지도록 이동
